Nvm Utility class functions can remain static problem was in namespaces
Chaged all the files back to using the utility as a static object/class. Which is much much better since it wont have to be recreated or passed everywhere. The problem with qt declaring the string codes in custom naqmespaces is still prevented since now the codes are a part of the utility class. This would make sense in a way since thats the only place they are used.