Disaster occurrence rates/fatalities research
"Population Research Sources.docx" contains the sources used for the rates of occurrence of disasters and the deaths per year per disaster, contained in "Rates of Population Change.pptx". The file contains many hyperlinks to websites from which the information is retrieved. The rate of occurrences and deaths per disaster aren't scientifically accurate, per se. They are as close to what the actual rates are as possible, but the point of the application is to simulate the world population, not simulating disaster occurrences.  Another thing is that the rates and number of deaths are probably rounded down; we didn't assume that other undocumented disasters happened, just those that are documented and somewhat verifiable.